make '+' button add a random King
